Move isSiteCustomizable to `state/selectors`
